vexa/services/bot-manager/requirements.txt
# Shared library dependency - REMOVED (Installed via Dockerfile RUN command)
# -e ../../libs/shared-models
fastapi
uvicorn[standard]
redis>=5.0.0
# docker>=6.0.0,<7.0.0 # REMOVED - No longer needed for counting
python-dotenv
# pydantic==1.10.7 # Now handled by shared-models
celery==5.3.0
# psycopg2-binary==2.9.6 # Re-add this for sync engine
# psycopg2-binary>=2.8 # Now handled by shared-models
# sqlalchemy>=2.0.0 # Now handled by shared-models
# asyncpg # Now handled by shared-models
# databases[postgresql]>=0.5.0 # Now handled by shared-models
email-validator # Added for Pydantic EmailStr support via shared-models
requests_unixsocket # Add for testing socket connection directly
# alembic # Optional: Add if database migrations are needed later

# Added for shared models/DB access:
SQLAlchemy[asyncio]
aiosqlite # If using sqlite async (often for testing)
# asyncpg # Alternative async PostgreSQL driver

# Added for bot limit check - REMOVED
# docker 

requests
httpx
psycopg2-binary
aiodocker
boto3  # MinIO/S3 storage for recordings
python-multipart  # Required for UploadFile/Form handling
deepgram-sdk>=3.0.0  # Required for batch file transcription